+++ packages/ts-lib-scripts/index.js
@@ -1,4 +1,3 @@
 #!/usr/bin/env node
-/* eslint-disable import/no-unresolved */
 
-require('./dist/index');
+require('./dist/index.js');