+++ packages/ts-lib-tools/src/config/jest/babelTransform.ts
@@ -1,6 +1,6 @@
 import babelJest from 'babel-jest';
 
-const babelTransform = babelJest.createTransformer({
+const babelTransform = babelJest.createTransformer?.({
   presets: [require.resolve('babel-preset-ts-lib')],
   babelrc: false,
   configFile: false,